Move filters out of events.ts
