+++ extension/src/events.ts
@@ -1,63 +1,20 @@
 import {
   Disposable,
-  DocumentSelector,
   Event,
-  languages,
   TextDocument,
   TextDocumentChangeEvent,
   workspace,
 } from "vscode";
 
 import {
-  excludeUriSchemes,
+  excludeIrrelevantChangeEventsByLanguage,
+  excludeIrrelevantChangeEventsByScheme,
+  excludeIrrelevantTextDocumentsByLanguage,
+  excludeIrrelevantTextDocumentsByScheme,
   ignoreIfAlreadyClosed,
-  select,
 } from "./events/filters";
 import { streamEvents } from "./events/stream";
 import { throttleEvent } from "./events/throttle";
-
-const schemesToExclude: string[] = ["git", "gitfs", "output", "vscode"];
-const ifmLanguageSelector: DocumentSelector = { language: "ifm" };
-
-function excludeIrrelevantTextDocumentsByLanguage(
-  event: Event<TextDocument>
-): Event<TextDocument> {
-  return select(
-    (document: TextDocument) =>
-      !!languages.match(ifmLanguageSelector, document),
-    event
-  );
-}
-
-function excludeIrrelevantChangeEventsByLanguage(
-  event: Event<TextDocumentChangeEvent>
-): Event<TextDocumentChangeEvent> {
-  return select(
-    (e: TextDocumentChangeEvent) =>
-      !!languages.match(ifmLanguageSelector, e.document),
-    event
-  );
-}
-
-function excludeIrrelevantTextDocumentsByScheme(
-  event: Event<TextDocument>
-): Event<TextDocument> {
-  return excludeUriSchemes(
-    schemesToExclude,
-    (document: TextDocument) => document.uri,
-    event
-  );
-}
-
-function excludeIrrelevantChangeEventsByScheme(
-  event: Event<TextDocumentChangeEvent>
-): Event<TextDocumentChangeEvent> {
-  return excludeUriSchemes(
-    schemesToExclude,
-    (e: TextDocumentChangeEvent) => e.document.uri,
-    event
-  );
-}
 
 const onDidInitiallyFindTextDocument: Event<TextDocument> = (
   listener: (e: TextDocument) => any,
